Fixed Jumping
This code handles touch events in an Android application. It limits the actions triggered by touch to occur only once every 0.9 milliseconds. When a touch event occurs, it checks if enough time has passed since the last touch. If so, it updates the game state to indicate a jump and sets the velocity of a square object. Then, it records the time of this touch for future reference. If the time between consecutive touches is less than 0.9, the touch event is ignored.

And changed the name of velocity when jumped to jumpVelocity and changed the velocity to -40 (faster)